Explicitly specify license in Cargo.toml
In the [docs for the Cargo manifest format](https://doc.rust-lang.org/cargo/reference/manifest.html#the-license-and-license-file-fields), the use of `license-file` is only recommended for non-standard licenses. For standard licenses such as MIT, it should be specified in the `license` field explicitly.

This change makes it easier for tools to correctly identify the license. For example, right now the license is identified as `non-standard` on both [crates.io](https://crates.io/crates/bevier) and the [Bevy Assets page](https://bevyengine.org/assets/#templates).